Summary: Fixes #33 - thread-executor
**Problem:**
Python's thread pool executor runs non-daemon threads.  If they time
out, there's no way to let the program abandon them.

This scenario fills up a thread pool, and blocks future threads from
doing work.

**Analysis:**
Create a thread executor which returns futures which can be abandoned,
and which can abandon all futures, allowing the main thread to exit if
the threads time out, or if the main thread doesn't need their output
anymore.

Just make it an executor instead of a pool.  We can wrap a pool around
it later if the thread startup cost becomes too high.

Specifically:
* thread executor submit works, holds program open as long as there are threads waiting.
* abandon functions work on both futures and the executor
* concurrent.futures.Executor - compliant executor for threads.
* added exit options for context manager.

**Testing:**
Added several unit tests, as well as several E2E examples (which are not
properly tests, but demonstrate use and have revealed interesting
problems that it's difficult to write unit tests for).

**Documentation:**
Added a bunch of in-line documentation to the functions and classes.

**Impact:**
Minor - added thread executor